+++ tests/unit/transformer/test_mapper.py
@@ -13,7 +13,7 @@ from spooq.transformer import Mapper
 
 @pytest.fixture(scope="module")
 def transformer(mapping):
-    return Mapper(mapping=mapping, ignore_missing_columns=True)
+    return Mapper(mapping=mapping, missing_column_handling="nullify")
 
 
 @pytest.fixture(scope="module")
@@ -135,12 +135,12 @@ class TestMultipleMappings(object):
 
     def test_appending_a_mapping(self, mapped_df, new_mapping, input_columns, new_columns):
         """Output schema is correct for added mapping at the end of the input schema"""
-        new_mapped_df = Mapper(mapping=new_mapping, mode="append", ignore_missing_columns=True).transform(mapped_df)
+        new_mapped_df = Mapper(mapping=new_mapping, mode="append", missing_column_handling="nullify").transform(mapped_df)
         assert input_columns + new_columns == new_mapped_df.columns
 
     def test_prepending_a_mapping(self, mapped_df, new_mapping, input_columns, new_columns):
         """Output schema is correct for added mapping at the beginning of the input schema"""
-        new_mapped_df = Mapper(mapping=new_mapping, mode="prepend", ignore_missing_columns=True).transform(mapped_df)
+        new_mapped_df = Mapper(mapping=new_mapping, mode="prepend", missing_column_handling="nullify").transform(mapped_df)
         assert new_columns + input_columns == new_mapped_df.columns
 
     def test_appending_a_mapping_with_duplicated_columns(self, input_columns, mapped_df):
@@ -152,7 +152,7 @@ class TestMultipleMappings(object):
         ]
         new_columns = [name for (name, path, data_type) in new_mapping]
         new_columns_deduplicated = [x for x in new_columns if x not in input_columns]
-        new_mapped_df = Mapper(mapping=new_mapping, mode="append", ignore_missing_columns=True).transform(mapped_df)
+        new_mapped_df = Mapper(mapping=new_mapping, mode="append", missing_column_handling="nullify").transform(mapped_df)
         assert input_columns + new_columns_deduplicated == new_mapped_df.columns
         assert mapped_df.schema["birthday"].dataType == T.TimestampType()
         assert new_mapped_df.schema["birthday"].dataType == T.DateType()
@@ -166,7 +166,7 @@ class TestMultipleMappings(object):
         ]
         new_columns = [name for (name, path, data_type) in new_mapping]
         new_columns_deduplicated = [x for x in new_columns if x not in input_columns]
-        new_mapped_df = Mapper(mapping=new_mapping, mode="prepend", ignore_missing_columns=True).transform(mapped_df)
+        new_mapped_df = Mapper(mapping=new_mapping, mode="prepend", missing_column_handling="nullify").transform(mapped_df)
         assert new_columns_deduplicated + input_columns == new_mapped_df.columns
         assert mapped_df.schema["birthday"].dataType == T.TimestampType()
         assert new_mapped_df.schema["birthday"].dataType == T.DateType()
@@ -179,7 +179,7 @@ class TestExceptionForMissingInputColumns(object):
 
     @pytest.fixture(scope="class")
     def transformer(self, mapping):
-        return Mapper(mapping=mapping, ignore_missing_columns=False, skip_missing_columns=False, nullify_missing_columns=False)
+        return Mapper(mapping=mapping, missing_column_handling="raise_error")
 
     def test_missing_column_raises_exception(self, input_df, transformer):
         input_df = input_df.drop("attributes")
@@ -199,7 +199,7 @@ class TestNullifyMissingColumns(object):
 
     @pytest.fixture(scope="class")
     def transformer(self, mapping):
-        return Mapper(mapping=mapping, skip_missing_columns=False, nullify_missing_columns=True)
+        return Mapper(mapping=mapping, missing_column_handling="nullify")
 
     @pytest.fixture(scope="class")
     def mapped_df(self, input_df, transformer):
@@ -222,7 +222,7 @@ class TestSkipMissingColumns(object):
 
     @pytest.fixture(scope="class")
     def transformer(self, mapping):
-        return Mapper(mapping=mapping, skip_missing_columns=True, nullify_missing_columns=False)
+        return Mapper(mapping=mapping, missing_column_handling="skip")
 
     @pytest.fixture(scope="class")
     def mapped_df(self, input_df, transformer):
@@ -234,14 +234,14 @@ class TestSkipMissingColumns(object):
         assert not any([column in mapped_df.columns for column in attribute_columns])
 
 
-class TestExceptionWhenBothSkippingAndNullifying(object):
+class TestExceptionWhenInvalidHandling(object):
     """
     Raise an exception in case both parameters skip_missing_columns and nullify_missing_columns are True
     """
 
     def test_invalid_parameter_setting_raises_exception(self, input_df, transformer):
         with pytest.raises(ValueError):
-            Mapper(mapping=mapping, skip_missing_columns=True, nullify_missing_columns=True)
+            Mapper(mapping=mapping, missing_column_handling="invalid")
 
 
 class TestDataTypesOfMappedDataFrame(object):